Add support for subscriptions and generic fragment list of user
